feat: added validation for task

+++ src/modules/task/commands/task/remove-task/remove-task.command-handler.ts
@@ -1,8 +1,10 @@
 import { ExceptionBase } from '@libs/base-classes';
 import { CommandHandlerBase } from '@libs/base-classes/command-handler.base';
+import { ConflictException } from '@libs/exceptions';
 import { Result } from '@libs/utils';
 import { UuidVO } from '@libs/value-objects';
 import { TaskUnitOfWork } from '@modules/task/database/unit-of-work';
+import { TaskStatus } from '@modules/task/types';
 import { CommandHandler } from '@nestjs/cqrs';
 
 import { RemoveTaskCommand } from './remove-task.command';
@@ -25,6 +27,15 @@ export class RemoveTaskCommandHandler extends CommandHandlerBase<
     } = command;
 
     const taskRepository = this.unitOfWork.getTaskRepository(trxId);
+
+    const taskResult = await taskRepository.getOneById(new UuidVO(id));
+    const task = taskResult.unwrap();
+
+    if (task.status === TaskStatus.COMPLETED) {
+      return Result.fail(
+        new ConflictException('Нельзя удалить выполненную задачу'),
+      );
+    }
 
     return taskRepository.remove(new UuidVO(id));
   }
